Address PR review findings on secret-store branch
- Consolidate dispatch: both adapters now resolve handles in run_app and
  call dispatch_with_handles directly; promotes resolve_kv_handle,
  resolve_secret_handle, dispatch_with_handles to pub(crate)
- Remove dead _secret_binding param from Cloudflare dispatch_with_kv_and_secrets
- Replace DummyKvStore with NoopKvStore in both adapter store_handles tests
- Extract shared EnvOverride + env_guard into axum test_utils module
- Export NoopKvStore under test-utils feature to match secret-store pattern
- Add doc comments to RequestContext::kv_handle and secret_handle
- Add GHAS safety comment to Secrets::from_request in extractor.rs
- Document breaking-change on run_app for both Fastly and Cloudflare adapters
