+++ crates/edgezero-adapter-axum/src/service.rs
@@ -98,41 +98,12 @@ impl Service<Request<AxumBody>> for EdgeZeroAxumService {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::test_utils::{env_guard, EnvOverride};
     use edgezero_core::body::Body;
     use edgezero_core::context::RequestContext;
     use edgezero_core::error::EdgeError;
     use edgezero_core::http::{response_builder, StatusCode};
-    use std::ffi::OsString;
-    use std::sync::OnceLock;
     use tower::ServiceExt;
-
-    fn env_guard() -> &'static tokio::sync::Mutex<()> {
-        static GUARD: OnceLock<tokio::sync::Mutex<()>> = OnceLock::new();
-        GUARD.get_or_init(|| tokio::sync::Mutex::new(()))
-    }
-
-    struct EnvOverride {
-        key: &'static str,
-        original: Option<OsString>,
-    }
-
-    impl EnvOverride {
-        fn set(key: &'static str, value: impl AsRef<std::ffi::OsStr>) -> Self {
-            let original = std::env::var_os(key);
-            std::env::set_var(key, value);
-            Self { key, original }
-        }
-    }
-
-    impl Drop for EnvOverride {
-        fn drop(&mut self) {
-            if let Some(ref original) = self.original {
-                std::env::set_var(self.key, original);
-            } else {
-                std::env::remove_var(self.key);
-            }
-        }
-    }
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     async fn forwards_request_to_router() {
