+++ crates/edgezero-core/src/lib.rs
@@ -19,6 +19,8 @@ pub mod router;
 pub mod secret_store;
 
 pub use edgezero_macros::{action, app};
+#[cfg(any(test, feature = "test-utils"))]
+pub use key_value_store::NoopKvStore;
 pub use key_value_store::{KvError, KvHandle, KvPage, KvStore};
 #[cfg(any(test, feature = "test-utils"))]
 pub use secret_store::{InMemorySecretStore, NoopSecretStore};
